% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/easyumap-package.R
\docType{package}
\name{easyumap-package}
\alias{easyumap}
\alias{easyumap-package}
\title{easyumap: What the Package Does (One Line, Title Case)}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sam El-Kamand \email{selkamand@ccia.org.au} (\href{https://orcid.org/0000-0003-2270-8088}{ORCID})

}
\keyword{internal}
