% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/umap_visualise.R
\name{umap_make_interactive}
\alias{umap_make_interactive}
\title{Make UMAP visualisation interactive}
\usage{
umap_make_interactive(
  gg,
  width_svg = NULL,
  height_svg = NULL,
  rescale = FALSE,
  width = 1
)
}
\arguments{
\item{gg}{the plot returned from any sigverse visualisation}

\item{width_svg, height_svg}{The width and height of the graphics region in inches.
The default values are 6 and 5 inches. This will define the aspect ratio of the
graphic as it will be used to define viewbox attribute of the SVG result.

If you use \code{girafe()} in an 'R Markdown' document, we
recommend not using these arguments so that the knitr
options \code{fig.width} and \code{fig.height} are used instead.}

\item{rescale}{If FALSE, graphic will not be resized and the dimensions are exactly
those of the svg. If TRUE the graphic will be resize to fit its container}

\item{width}{widget width ratio (0 < width <= 1).}
}
\value{
a ggiraph interactive visualisation
}
\description{
Make UMAP visualisation interactive
}
